- Rename of input from 'include_directories' to 'paths'. - Removed support for case sensitivity and suffix vs contains after introducing glob. - Rename include_directories to paths.
